Bumeet/pyproject.toml
[tool.poetry]
name = "bumeet"
version = "0.1.0"
description = ""
authors = ["ormor <htth815@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
django = "5.0"
djangorestframework = "3.15.2"
pillow = "10.4.0"
langchain-community = "0.3.1"
tiktoken = "^0.8.0"
langchain-openai = "0.2.1"
langchainhub = "^0.1.21"
chromadb = "^0.5.11"
python-dotenv = "^1.0.1"
langsmith = "^0.1.131"
tavily-python = "^0.5.0"
langgraph = "^0.2.34"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
